Refactor display and firework effect classes; consolidate color handling and removed deprecated code
